Move xml.dom._types to _typeshed.xml
