+++ stdlib/2and3/_typeshed/xml.pyi
